test: WinningPrizeTest 다른곳에 썼던 것 옮김
